HxAutosuggest & derived components cleanup - HxXyInternal components naming
