+++ Havit.Blazor.Components.Web.Bootstrap/Forms/Autosuggests/Internal/HxAutosuggestInputInternal.razor
@@ -11,9 +11,9 @@
        value="@Value"
        @oninput="HandleInput"
        @onkeydown="OnKeyDown"
-       @onfocus="OnInputFocus"
-       @onblur="OnInputBlur"
-       @onmousedown="OnInputMouseDown"
+       @onfocus="OnFocus"
+       @onblur="OnBlur"
+       @onmousedown="OnMouseDown"
        @onclick:stopPropagation
        placeholder="@Placeholder"
        onfocusin="this.select()"
