--- Havit.Blazor.Components.Web.Bootstrap/Forms/SearchBox/Internal/HxSearchBoxItem.razor.cs
@@ -1,14 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-
-namespace Havit.Blazor.Components.Web.Bootstrap.Internal;
-
-public partial class HxSearchBoxItem
-{
-	[Parameter] public string Title { get; set; }
-	[Parameter] public string Subtitle { get; set; }
-	[Parameter] public IconBase Icon { get; set; }
-}
