+++ Havit.Blazor.Components.Web.Bootstrap/Forms/SearchBox/Internal/HxSearchBoxItemInternal.razor.cs
@@ -1,0 +1,8 @@
+﻿namespace Havit.Blazor.Components.Web.Bootstrap.Internal;
+
+public partial class HxSearchBoxItemInternal
+{
+	[Parameter] public string Title { get; set; }
+	[Parameter] public string Subtitle { get; set; }
+	[Parameter] public IconBase Icon { get; set; }
+}
