+++ Havit.Blazor.Components.Web.Bootstrap/Tags/Internal/HxInputTagsAutosuggestInputInternal.razor.cs
@@ -1,18 +1,18 @@
 ﻿namespace Havit.Blazor.Components.Web.Bootstrap.Internal
 {
-	public partial class HxAutosuggestInput
+	public partial class HxInputTagsAutosuggestInputInternal
 	{
 		[Parameter] public string Value { get; set; }
 
 		[Parameter] public string Placeholder { get; set; }
 
-		[Parameter] public EventCallback<string> OnInputInput { get; set; }
+		[Parameter] public EventCallback<string> OnInput { get; set; }
 
-		[Parameter] public EventCallback OnInputFocus { get; set; }
+		[Parameter] public EventCallback OnFocus { get; set; }
 
-		[Parameter] public EventCallback OnInputBlur { get; set; }
+		[Parameter] public EventCallback OnBlur { get; set; }
 
-		[Parameter] public EventCallback OnInputMouseDown { get; set; }
+		[Parameter] public EventCallback OnMouseDown { get; set; }
 
 		[Parameter] public EventCallback<KeyboardEventArgs> OnKeyDown { get; set; }
 
@@ -23,21 +23,17 @@
 		[Parameter] public bool EnabledEffective { get; set; }
 
 		/// <summary>
-		/// Offset between the dropdown and the input.
-		/// <see href="https://popper.js.org/docs/v2/modifiers/offset/#options"/>
+		/// Offset between dropdown input and dropdown menu
 		/// </summary>
-		[Parameter] public (int Skidding, int Distance) DropdownOffset { get; set; }
+		[Parameter] public (int X, int Y) Offset { get; set; }
 
-		/// <summary>
-		/// Additional attributes to be splatted onto an underlying HTML element.
-		/// </summary>
 		[Parameter(CaptureUnmatchedValues = true)] public Dictionary<string, object> AdditionalAttributes { get; set; }
 
 		internal ElementReference InputElement { get; set; }
 
 		private async Task HandleInput(ChangeEventArgs changeEventArgs)
 		{
-			await OnInputInput.InvokeAsync((string)changeEventArgs.Value);
+			await OnInput.InvokeAsync((string)changeEventArgs.Value);
 		}
 
 		public async ValueTask FocusAsync()
